Add RAW mode patch and error log
Adapts the RAW mode pull request from the main repo so that external
processes can decode and simply pass the audio frames to the
transmitter. Also, adds the ability to pass num_channels and the
sample_rate so that one can play all manner of audio types so long as
they know the configuration. Finally, in the event of an error during
loading of input, sndfile logging is printed to stderr.