Still tweaking numbers with no scientific basis for the changes
I did notice on a radio that actually indicates whether or not the
signal is stereo that said changes weren't appearing as stereo.
Possibilities are either that the modulation still isn't correct, or
more simply, that the antenna I'm using is insufficient or badly placed,
as the RDS text was also sporadic.